+++ Analyzers/CQTAnalyzer.cpp
@@ -12,7 +12,7 @@
 /// <summary>
 /// Initializes a new instance.
 /// </summary>
-CQTAnalyzer::CQTAnalyzer(const Configuration * configuration, double sampleRate, uint32_t channelCount, uint32_t channelSetup, const WindowFunction & windowFunction) : Analyzer(configuration, sampleRate, channelCount, channelSetup, windowFunction)
+CQTAnalyzer::CQTAnalyzer(const Configuration * configuration, uint32_t sampleRate, uint32_t channelCount, uint32_t channelSetup, const WindowFunction & windowFunction) : Analyzer(configuration, sampleRate, channelCount, channelSetup, windowFunction)
 {
 }
 
@@ -23,19 +23,19 @@ bool CQTAnalyzer::AnalyzeSamples(const audio_sample * sampleData, size_t sampleC
 {
     for (FrequencyBand & Iter : frequencyBands)
     {
-        double Bandwidth = ::fabs(Iter.Hi - Iter.Lo) + (_SampleRate / (double) sampleCount) * _Configuration->_CQTBandwidthOffset;
-        double TLen = Min(1. / Bandwidth, (double) sampleCount / _SampleRate);
+        double Bandwidth = ::fabs(Iter.Hi - Iter.Lo) + ((double) _SampleRate / (double) sampleCount) * _Configuration->_CQTBandwidthOffset;
+        double TLen = Min(1. / Bandwidth, (double) sampleCount / (double) _SampleRate);
 
-        double DownsampleAmount = Max(1.0, ::trunc((_SampleRate * _Configuration->_CQTDownSample) / (Iter.Ctr + TLen)));
-        double Coeff = 2. * ::cos(2. * M_PI * Iter.Ctr / _SampleRate * DownsampleAmount);
+        double DownsampleAmount = Max(1.0, ::trunc(((double) _SampleRate * _Configuration->_CQTDownSample) / (Iter.Ctr + TLen)));
+        double Coeff = 2. * ::cos(2. * M_PI * Iter.Ctr / (double) _SampleRate * DownsampleAmount);
 
         double f1 = 0.;
         double f2 = 0.;
         double Sine = 0.;
-        double Offset = ::trunc(((double) sampleCount - TLen * _SampleRate) * (0.5 + _Configuration->_CQTAlignment / 2.));
+        double Offset = ::trunc(((double) sampleCount - TLen * (double) _SampleRate) * (0.5 + _Configuration->_CQTAlignment / 2.));
 
         double LoIdx = Offset;
-        double HiIdx = ::trunc(TLen * _SampleRate) + Offset - 1.;
+        double HiIdx = ::trunc(TLen * (double) _SampleRate) + Offset - 1.;
         double Norm = 0.;
 
         for (double Idx = ::trunc(LoIdx / DownsampleAmount); Idx <= ::trunc(HiIdx / DownsampleAmount); ++Idx)
