+++ Analyzers/CQTAnalyzer.h
@@ -26,6 +26,6 @@ public:
 
     virtual ~CQTAnalyzer() { }
 
-    CQTAnalyzer(const Configuration * configuration, double sampleRate, uint32_t channelCount, uint32_t channelSetup, const WindowFunction & windowFunction);
+    CQTAnalyzer(const Configuration * configuration, uint32_t sampleRate, uint32_t channelCount, uint32_t channelSetup, const WindowFunction & windowFunction);
     bool AnalyzeSamples(const audio_sample * sampleData, size_t sampleCount, vector<FrequencyBand> & frequencyBands) const;
 };
